Move conversions into own module
