+++ pyrs/utilities/checkdatatypes.py
@@ -1,4 +1,5 @@
 # PyRS static helper methods
+from .convertdatatypes import to_float
 import os
 import numpy
 import six
@@ -97,18 +98,8 @@ def check_float_variable(var_name, variable, value_range):
     :except ValueError:
     :param value_range: if not None, then must be a 2 tuple as [min, max)
     '''
-    check_string_variable('var_name', var_name)
-    assert isinstance(variable, (float, int, numpy.integer, numpy.floating)), '{0} {1} must be a float but not a {2}'\
-        .format(var_name, variable, type(variable))
-
-    if value_range is not None:
-        assert len(value_range) == 2, '{0}\' value range {1} must be either a None or have 2 elements as [min, max)' \
-                                      ''.format(var_name, value_range)
-
-        min_val = value_range[0]
-        max_val = value_range[1]
-        if (min_val is not None and variable < min_val) or (max_val is not None and variable >= max_val):
-            raise ValueError('{0} (= {1}) is out of range [{2}, {3})'.format(var_name, variable, min_val, max_val))
+    # use newer code and ignore return
+    _ = to_float(var_name, variable, value_range[0], value_range[1])
 
 
 def check_list(var_name, variable, allowed_values=None):
